+++ src/utils/setup.js
@@ -1,6 +1,10 @@
 import Relay from 'react-relay';
 import { serverURL } from '../../env';
 
+import * as snapshot from './snapshot';
+
+// snapshot.clearSnapshot();
+
 Relay.injectNetworkLayer(
   new Relay.DefaultNetworkLayer(`${serverURL}/graphql`, {
     fetchTimeout: 30000,